requirements.txt
faster-whisper
langchain-core
langchain-openai
openai
pyaudio
python-dotenv
rich
speechrecognition
